+++ backends/default/xla.go
@@ -1,4 +1,7 @@
-//go:build !noxla
+//go:build linux && amd64 && !noxla
+
+// For now XLA is only supported for linux/amd64.
+// TODO: change when more platforms are supported (linux/arm64, darwin/amd64, darwin/arm64, etc.)
 
 package _default